reenable logging and rename test file
