+++ cluster_orchestrator/IDP-docker-compose.yml
@@ -87,7 +87,7 @@ services:
       - MQTT_BROKER_PORT=10003
       - MQTT_BROKER_URL=mqtt
       - ROOT_SERVICE_MANAGER_URL=192.168.178.22
-      - ROOT_SERVICE_MANAGER_URL_v6=[2001:a61:1254:e301:b3a7:f62e:f4b0:e291]
+      - ROOT_SERVICE_MANAGER_URL_v6=[2001:a61:12bc:8301:aad4:89b7:65b6:5f76]
       - ROOT_SERVICE_MANAGER_PORT=10099
       - SYSTEM_MANAGER_URL=cluster_manager
       - SYSTEM_MANAGER_PORT=10000
@@ -122,7 +122,7 @@ services:
     environment:
       - MY_PORT=10100
       - SYSTEM_MANAGER_URL=192.168.178.22
-      - SYSTEM_MANAGER_URL_v6=[2001:a61:1254:e301:b3a7:f62e:f4b0:e291]
+      - SYSTEM_MANAGER_URL_v6=[2001:a61:12bc:8301:aad4:89b7:65b6:5f76]
       - SYSTEM_MANAGER_PORT=10000
       - CLUSTER_SERVICE_MANAGER_ADDR=cluster_service_manager
       - CLUSTER_SERVICE_MANAGER_PORT=10110
